+++ lib/generators/recording_studio_support/install/templates/recording_studio_support_initializer.rb
@@ -2,6 +2,7 @@
 
 RecordingStudioSupport.configure do |config|
   config.pages_path = "/support"
+  config.public_pages_path = "/help"
   config.pages_title = "Help"
   config.pages_subtitle = "Answers you can share."
   config.admin_section_title = "Help"