+++ test/dummy/app/helpers/application_helper.rb
@@ -1,2 +1,5 @@
 module ApplicationHelper
+  include RecordingStudioSupport::ApplicationHelper
+  include RecordingStudioPublishable::ApplicationHelper
 end
+