+++ test/dummy/config/initializers/recording_studio_support.rb
@@ -2,8 +2,10 @@
 
 RecordingStudioSupport.configure do |config|
   config.pages_path = "/support"
+  config.public_pages_path = "/help"
   config.pages_title = "Help"
   config.pages_subtitle = "Answers you can share."
   config.admin_section_title = "Help"
   config.admin_section_subtitle = "Pages people use when they get stuck."
 end
+